Move SourceHabitat to habitat_models.py
